requirements.txt
pandas==2.2.2
numpy>=1.26.0
torch>=2.2.0
transformers>=4.43.0
accelerate>=0.33.0
scikit-learn>=1.4.0
pydantic>=2.7.0
PyYAML>=6.0
python-crfsuite>=0.9.9
tqdm>=4.66.0
